feat(institution): add institution management to routes
